Added more Tests. Initial Quaternion.

+++ benchmarks/vector_operator.cpp
@@ -3,7 +3,7 @@
 
 void benchAddOperator(benchmark::State& state)
 {
-    xme::Vector<double, 4> v{1, 2, 3, 4};
+    xme::Vector<float, 4> v{1, 2, 3, 4};
     for(auto _ : state) {
         auto r = v * 5;
         benchmark::DoNotOptimize(r);
